% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{geom_lime}
\alias{geom_lime}
\title{Draw limes}
\usage{
geom_lime(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
documentation for \code{ggplot2::geom_point()}}
}
\description{
Draw limes
}
